feat: updated elash to use new UI
Converted the DMs system to use the new discord UI components, and
created a `UserUniqueView` to help tailor messages to single users in a
shared channel. This is quite limited unfortunately, as I can't find in
the docs a way to send an ephemeral message in a channel without going
through the `interaction.response.send_message` method.

As such, the user is asked to click a button to get their prompt, which
is a suitable workaround I suppose.

The modal input + safeties is a little funky, so we know determine the
safeties before asking for answers so we can actually display what the
safety is. Consequently the UI components return both the message (which
may be the default) and a flag to indicate if the safety was used or
not.

Needs a lot of polish. The interaction messages can be embeds too, so we
can get pretty text formatting.

Not extensively tested.