feat: emoji hell
Added a big list of emojis and a random emoji function. This is used to
assign a random emoji to each player, and to all of the UI buttons for a
bit more flair.

Downside of this is that we can no longer use the button decorators, as
the random emoji function would only be called during module load. Moved
all relevant buttons into the body.

Added a (useless?) helper function for context wrapping async functions
to pass the `self` argument. It seems Python already does that when
qualifying with `self.function_name`?