Feat: new UI components (#52)
* chore: added pyproject.toml

Migrating to use pdm instead of poetry or pipenv.

* chore: add runner script

* feat: replaced player gathering dialog

Updated the player gathering dialog with the new UI component features.
Implemented a `TimedView` class as a helper for creating timeout views
that display the time remaining in the embed.

Needs polish.

* chore: add pyproject scripts for lint and more

* feat: updated elash to use new UI

Converted the DMs system to use the new discord UI components, and
created a `UserUniqueView` to help tailor messages to single users in a
shared channel. This is quite limited unfortunately, as I can't find in
the docs a way to send an ephemeral message in a channel without going
through the `interaction.response.send_message` method.

As such, the user is asked to click a button to get their prompt, which
is a suitable workaround I suppose.

The modal input + safeties is a little funky, so we know determine the
safeties before asking for answers so we can actually display what the
safety is. Consequently the UI components return both the message (which
may be the default) and a flag to indicate if the safety was used or
not.

Needs a lot of polish. The interaction messages can be embeds too, so we
can get pretty text formatting.

Not extensively tested.

* feat: ui fixes and poll

Added a `PollView` with the new UI components.

Updated elash to use new `PollView` component.

Intermediary results now delete themselves after some timeout.

Fixed bugs related to being able to vote and enter two results by
whitelisting which user IDs may interact and checking whether a user has
already interacted.

Fixed bugs related to getting input from modal text inputs when the
label text was >45 characters by hardcoding the label to "Answer".

Timed UI components now skip the timer when everyone who is expected to
have interacted has indeed interacted.

* fix: revert timeout for gathering and min players in elash

* fix: increase prompt time and vote tallies

Bumped the time to submit prompts from 15 seconds to 30.

Show the vote tallies in the final vote screen as the label of the embed
field.

* feat: emoji hell

Added a big list of emojis and a random emoji function. This is used to
assign a random emoji to each player, and to all of the UI buttons for a
bit more flair.

Downside of this is that we can no longer use the button decorators, as
the random emoji function would only be called during module load. Moved
all relevant buttons into the body.

Added a (useless?) helper function for context wrapping async functions
to pass the `self` argument. It seems Python already does that when
qualifying with `self.function_name`?

* fix: emoji mapping can't fail

There are some string codes that don't map to emojis, so we just keep
trying to convert until we get one. Bad strings get logged as warnings
so I can maybe one day grep through the logs and remove them.

* style: formatting

* fix: timeouts for elash rounds and safety tally

Fixed logic error when deducing which players used safeties. We now
instead just track whether a reply was a safety or not with a boolean
attached to the reply itself.

Timeout for votes and rounds increased to 1 minute, from 30 seconds.

* chore: update requirements